final version by thocle

+++ PROJET-ASSO/delete.php
@@ -1,25 +1,15 @@
 <?php
-    function loadClass($class)
-    {
-        if (str_contains($class, "Manager")) {
-            require "./controllers/$class.php";
-        } else {
-            require "./models/$class.php";
-        }
-    }
-    spl_autoload_register("loadClass");
-    $manager = new ArticleManager();
+session_start();
+require("./controllers/ArticleManager.php");
+// require("./controllers/UserManager.php");
+
+$manager = new ArticleManager();
+if ($_SESSION['status'] == True) {
     if ($_GET) {
         $manager->delete($_GET['id']);
     }
-
-    if (str_contains($_SERVER['HTTP_REFERER'],"readAll") == TRUE ) {
-        echo "<script> window.location.href='readall.php'</script>";
-    }else if (str_contains($_SERVER['HTTP_REFERER'],"top_article") == TRUE ){
-        echo "<script> window.location.href='top_article.php'</script>";
-    }else{
-        echo "<script> window.location.href='readall.php'</script>";
-    }
-
-    ?>
-    
+}
+else {
+    echo "<script> alert('you must be admin to do that') </script>";
+}
+echo "<script> window.location.href='index.php'</script>";
